+++ src/Fields/index.ts
@@ -13,7 +13,7 @@ import Text from './Text';
 import Field from './Field';
 import Enum from './Enum';
 import List from './List';
-import Float from './Float';
+import Decimal from './Decimal';
 
 export {
   Array,
@@ -31,5 +31,5 @@ export {
   Field,
   Enum,
   List,
-  Float,
+  Decimal,
 };
